Remove reflection, fix breakage
There was (and is) still breakage due to the switch from com.ning to
org.asynchttpclient. Squashing reflection warnings was helpful in
finding places where interop calls needed to be changed.

There were a few places where there was no obvious equivalent in the new
java code, so I had to break a few things.

Breakage:

- create-client now takes `thread-factory` instead of `executor-service`
- conversion of response headers to a map is no longer lazy. It could
  be made lazy again if there really is value in that, but I just kept
  it simple and built a plain-old clojure map instead.
- the `auth` map no longer supports a `:target-proxy` key, since the
  setTargetProxy method seems to be gone, and I don't know what (if
  anything) is the new equivalent.